Bugfix: dont delete currently used profiles/shared folders
git-svn-id: file:///home/code/svn/ovd/trunk@5312 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/admin/actions.php
@@ -1040,7 +1040,10 @@ if ($_REQUEST['name'] == 'NetworkFolders') {
 		foreach ($_REQUEST['ids'] as $id) {
 			$network_folder = Abstract_NetworkFolder::load($id);
 			if (is_object($network_folder))
-				$buf = Abstract_NetworkFolder::delete($network_folder);
+				if (! $network_folder->isUsed())
+					$buf = Abstract_NetworkFolder::delete($network_folder);
+				else
+					$buf = false;
 
 			if (! $buf)
 				popup_error(sprintf(_("Unable to delete network folder '%s'"), $network_folder->name));
